add platform to Arduino
