+++ apps/microtvm/arduino/template_project/src/host_driven/platform.c
@@ -17,42 +17,38 @@
  * under the License.
  */
 
-#include "model.h"
-
-#include "Arduino.h"
 #include "standalone_crt/include/dlpack/dlpack.h"
-#include "standalone_crt/include/tvm/runtime/crt/stack_allocator.h"
+#include "standalone_crt/include/tvm/runtime/crt/error_codes.h"
+#include "stdarg.h"
 
-// AOT memory array, stack allocator wants it aligned
-static uint8_t g_aot_memory[WORKSPACE_SIZE]
-    __attribute__((aligned(TVM_RUNTIME_ALLOC_ALIGNMENT_BYTES)));
-tvm_workspace_t app_workspace;
+// Called by TVM when a message needs to be formatted.
+__attribute__((weak)) size_t TVMPlatformFormatMessage(char* out_buf, size_t out_buf_size_bytes,
+                                                      const char* fmt, va_list args) {
+  return vsnprintf(out_buf, out_buf_size_bytes, fmt, args);
+}
 
+// Called by TVM when an internal invariant is violated, and execution cannot continue.
 // Blink code for debugging purposes
-void TVMPlatformAbort(tvm_crt_error_t error) {
+__attribute__((weak)) void TVMPlatformAbort(tvm_crt_error_t error) {
   TVMLogf("TVMPlatformAbort: 0x%08x\n", error);
-  for (;;) {
-#ifdef LED_BUILTIN
-    digitalWrite(LED_BUILTIN, HIGH);
-    delay(250);
-    digitalWrite(LED_BUILTIN, LOW);
-    delay(250);
-    digitalWrite(LED_BUILTIN, HIGH);
-    delay(250);
-    digitalWrite(LED_BUILTIN, LOW);
-    delay(750);
-#endif
+  for (;;)
+    ;
+}
+
+// Called by TVM when memory allocation is required.
+__attribute__((weak)) tvm_crt_error_t TVMPlatformMemoryAllocate(size_t num_bytes, DLDevice dev,
+                                                                void** out_ptr) {
+  if (num_bytes == 0) {
+    num_bytes = sizeof(int);
   }
+  *out_ptr = malloc(num_bytes);
+  return (*out_ptr == NULL) ? kTvmErrorPlatformNoMemory : kTvmErrorNoError;
 }
 
-void TVMLogf(const char* msg, ...) {}
-
-tvm_crt_error_t TVMPlatformMemoryAllocate(size_t num_bytes, DLDevice dev, void** out_ptr) {
-  return StackMemoryManager_Allocate(&app_workspace, num_bytes, out_ptr);
-}
-
-tvm_crt_error_t TVMPlatformMemoryFree(void* ptr, DLDevice dev) {
-  return StackMemoryManager_Free(&app_workspace, ptr);
+// Called by TVM to free an allocated memory.
+__attribute__((weak)) tvm_crt_error_t TVMPlatformMemoryFree(void* ptr, DLDevice dev) {
+  free(ptr);
+  return kTvmErrorNoError;
 }
 
 unsigned long g_utvm_start_time_micros;
@@ -82,13 +78,4 @@ tvm_crt_error_t TVMPlatformGenerateRandom(uint8_t* buffer, size_t num_bytes) {
     buffer[i] = rand();
   }
   return kTvmErrorNoError;
-}
-
-void TVMInitialize() { StackMemoryManager_Init(&app_workspace, g_aot_memory, WORKSPACE_SIZE); }
-
-void TVMExecute(void* input_data, void* output_data) {
-  int ret_val = tvmgen_default___tvm_main__(input_data, output_data);
-  if (ret_val != 0) {
-    TVMPlatformAbort(kTvmErrorPlatformCheckFailure);
-  }
 }
